+++ changelog.md
@@ -1,6 +1,8 @@
 # Unreleased
 
-- 
+- Add %player% variable for messages on button
+- Add PlaceholderAPI support for custom model data (modelID) 
+- Add argument for command /zm open
 
 # 1.0.0.8